cl: split composite panic regression for bare metal
