+++ cl/_testgo/complitassign/in.go
@@ -7,7 +7,6 @@ type pair struct {
 }
 
 var state pair
-var calls int
 
 // The right-hand side must be completely evaluated before state is replaced.
 // CHECK-LABEL: define void @main.assignGlobal(){{.*}} {
@@ -33,32 +32,7 @@ func second() int {
 	return 2
 }
 
-// Both calls must run before assigning through the nil pointer panics.
-// CHECK-LABEL: define void @main.assignNil(){{.*}} {
-// CHECK: call i64 @main.count()
-// CHECK: call i64 @main.count()
-// CHECK: call void @"{{.*}}/runtime/internal/runtime.AssertNilDeref"
-func assignNil() {
-	defer func() {
-		if recover() == nil {
-			panic("nil composite assignment did not panic")
-		}
-		if calls != 2 {
-			panic("nil composite assignment panicked before evaluating its right-hand side")
-		}
-	}()
-
-	var dst *pair
-	*dst = pair{count(), count()}
-}
-
-func count() int {
-	calls++
-	return calls
-}
-
 func main() {
 	assignGlobal()
-	assignNil()
 	println("ok")
 }
