+++ cl/_testgo/complitnil/expect.txt
@@ -1,0 +1,1 @@
+ok
